@@ -1,8 +1,6 @@
-'use strict';
-
-const assert = require('assert'),
-    fs = require('fs-extra'),
-    formatter = require('../src/default');
+import assert from 'assert';
+import { writeFileSync, unlinkSync } from 'node:fs';
+import formatter from '../src/default.js';
 
 describe('default', function () {
     describe('#load', function () {
@@ -14,10 +12,10 @@ describe('default', function () {
                     name: 'name'
                 }]
             };
-            fs.writeFileSync('defaultTest.json', JSON.stringify(config));
+            writeFileSync('defaultTest.json', JSON.stringify(config));
 
             assert.deepStrictEqual(formatter.load({ configfile: 'defaultTest.json' }), config);
-            fs.unlinkSync('defaultTest.json');
+            unlinkSync('defaultTest.json');
         });
 
         it('should add imposters array if it is missing', function () {
@@ -26,10 +24,10 @@ describe('default', function () {
                 protocol: 'test',
                 name: 'name'
             };
-            fs.writeFileSync('defaultTest.json', JSON.stringify(config));
+            writeFileSync('defaultTest.json', JSON.stringify(config));
 
             assert.deepStrictEqual(formatter.load({ configfile: 'defaultTest.json' }), { imposters: [config] });
-            fs.unlinkSync('defaultTest.json');
+            unlinkSync('defaultTest.json');
         });
 
         it('should interpret EJS code blocks', function () {
@@ -110,7 +108,7 @@ describe('default', function () {
             formatter.save({ savefile: 'defaultTest.json' }, config);
 
             assert.deepStrictEqual(formatter.load({ configfile: 'defaultTest.json' }), config);
-            fs.unlinkSync('defaultTest.json');
+            unlinkSync('defaultTest.json');
         });
     });
 });
