@@ -1,13 +1,12 @@
-'use strict';
+import assert from 'assert';
+import { unlinkSync, writeFileSync } from 'node:fs';
+import formatter from '../src/noParse.js';
 
-const assert = require('assert'),
-    fs = require('fs-extra'),
-    formatter = require('../src/noParse'),
-    filename = 'noParseTest.json';
+const filename = 'noParseTest.json';
 
 describe('noParse', function () {
     afterEach(function () {
-        fs.unlinkSync(filename);
+        unlinkSync(filename);
     });
 
     describe('#load', function () {
@@ -19,7 +18,7 @@ describe('noParse', function () {
                     name: '<% include template.ejs %>'
                 }]
             };
-            fs.writeFileSync(filename, JSON.stringify(config));
+            writeFileSync(filename, JSON.stringify(config));
 
             assert.deepStrictEqual(formatter.load({ configfile: filename }), config);
         });
@@ -30,7 +29,7 @@ describe('noParse', function () {
                 protocol: 'test',
                 name: '<% include template.ejs %>'
             };
-            fs.writeFileSync(filename, JSON.stringify(config));
+            writeFileSync(filename, JSON.stringify(config));
 
             assert.deepStrictEqual(formatter.load({ configfile: filename }), { imposters: [config] });
         });
